+++ VMRay/vmray_modules/submit_url_action.py
@@ -2,7 +2,7 @@
 
 from vmray_modules.base import VMRayAction
 from vmray_modules.models import SubmitResults, SubmitUrlArguments
-from vmray_modules.submit_helpers import parse_submit_response
+from vmray_modules.submit_helpers import parse_submit_response, submission_params
 
 
 class SubmitUrl(VMRayAction):
@@ -16,5 +16,6 @@ class SubmitUrl(VMRayAction):
             tags=arguments.tags,
             reanalyze=arguments.reanalyze,
             analysis_caching=arguments.analysis_caching,
+            **submission_params(arguments),
         )
         return parse_submit_response(response)